requirements.txt
langchain == 0.1.16
langchain-openai==0.1.3
faiss-cpu==1.8.0
pandas==2.2.2
requests == 2.31.0
beautifulsoup4 == 4.12.3
tqdm==4.66.4
python-dotenv==1.0.1
